split webcontroller into 2 controllers
