+++ src/main/java/hello/controllers/HomeController.java
@@ -1,5 +1,6 @@
-package hello;
+package hello.controllers;
 
+import hello.EarthquakeQueryService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.oauth2.client.authentication.OAuth2AuthenticationToken;
 import org.springframework.security.oauth2.client.registration.ClientRegistration;
@@ -16,7 +17,7 @@ import com.nimbusds.oauth2.sdk.client.ClientReadRequest;
 import hello.geojson.FeatureCollection;
 
 @Controller
-public class WebController {
+public class HomeController {
 
     @Autowired
     private ClientRegistrationRepository clientRegistrationRepository;
@@ -38,35 +39,5 @@ public class WebController {
 
         model.addAttribute("urls", urls);
         return "login";
-    }
-
-    @GetMapping("/page1")
-    public String getPage1(Model model, OAuth2AuthenticationToken oAuth2AuthenticationToken) {
-
-        return "page1";
-    }
-
-    @GetMapping("/page2")
-    public String getPage2(Model model, OAuth2AuthenticationToken oAuth2AuthenticationToken) {
-
-        return "page2";
-    }
-
-    @GetMapping("/earthquakes/search")
-    public String getEarthquakesSearch(Model model, OAuth2AuthenticationToken oAuth2AuthenticationToken,
-            EqSearch eqSearch) {
-        return "earthquakes/search";
-    }
-
-    @GetMapping("/earthquakes/results")
-    public String getEarthquakesResults(Model model, OAuth2AuthenticationToken oAuth2AuthenticationToken,
-          EqSearch eqSearch) {
-          EarthquakeQueryService e = new EarthquakeQueryService();
-          model.addAttribute("eqSearch", eqSearch);
-          String json = e.getJSON(eqSearch.getDistance(), eqSearch.getMinmag());
-          model.addAttribute("json", json);
-          FeatureCollection featureCollection = FeatureCollection.fromJSON(json);
-          model.addAttribute("featureCollection",featureCollection);
-          return "earthquakes/results";
     }
 }
